4x4 matrix keypad input to microbit
basic set up to use a 4x4 matrix keypad with a bbc microbit, the LED matrix will display the number being pressed. Uses Pins 1,8,11,12 for the vertical axis and 13,14,15,16 for the horizontal axis, see the excellent explaination of how a matrix keypad works at http://pcbheaven.com/wikipages/How_Key_Matrices_Works/ to assist with understanding this set up. Very much a chance to make this DRYer if anyone's interested.